+++ src/lib/components/EmptyEpisodes.svelte
@@ -1,0 +1,25 @@
+<script lang="ts">
+	import EpisodeCore from './EpisodeCore.svelte';
+	let { episodesCount = 3 } = $props();
+</script>
+
+<div class="my-8 grid gap-6">
+	<!-- TODO: this should be handled a level up and then a flip animation can be used. -->
+	{#each Array(episodesCount) as _}
+		<article class="flex animate-pulse flex-col gap-4 rounded-xl border border-border">
+			<EpisodeCore
+				episode={{
+					title: '',
+					attachments: [],
+					content_html: '',
+					image: '',
+					authors: [],
+					date_published: '',
+					id: '',
+					url: '',
+					summary: ''
+				}}
+			/>
+		</article>
+	{/each}
+</div>
